backend/pyproject.toml
[project]
name = "modular-rag-mvp-backend"
version = "0.1.0"
description = "FastAPI backend for Modular RAG MVP"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi[standard]==0.115.0",
  "uvicorn[standard]==0.30.6",
  "pydantic==1.10.18",
  "python-dotenv==1.0.1",
  "requests==2.32.3",
  "numpy==1.26.4",
  "google-generativeai>=0.8.5",
]

[project.optional-dependencies]
faiss = [
  # choose one depending on platform
  # "faiss-cpu>=1.7.4",
]
embeddings = [
  # "sentence-transformers>=2.7.0",
  # or FlagEmbedding for BGE
  # "FlagEmbedding>=1.2.10"
]
openai = [
  # "openai>=1.52.0"
]
gemini = [
  # "google-generativeai>=0.8.3"
]

[tool.uv]
dev-dependencies = []


